Remove auth from blueprint and rename auth.py as decorators.py
