+++ app/__init__.py
@@ -17,10 +17,8 @@ def create_app(config_class=Config):
     migrate.init_app(app, db)
     ma.init_app(app)
 
-    from app.auth import auth_bp
     from app.main import main_bp
 
-    app.register_blueprint(auth_bp)
     app.register_blueprint(main_bp)
 
     return app
